Removed unnecessary files, cleaned UI.js, minor style changes
Removed any unnecessary files that are not currently being used, may be added later on if my plan of execution is more apparent. Cleaned up the UI-Display.js file so that the code is less cluttered with functions, .map and .join. Added the state or localized  equivalent to appear alongside the city. adjusted the background color to adapt if its currently day or night. minor css additions.